requirements.txt
discord
python-dotenv
asyncio
sqlalchemy
aiosqlite
aiohttp
tweepy
opencv-python
anthropic